\name{mMRF-package}
\alias{mMRF-package}
\alias{mMRF}
\docType{package}
\title{
Estimate Markov random fields
}
\description{
Estimates Markov random fields with a L1 constraint. 
}
\details{
\tabular{ll}{
Package: \tab mMRF\cr
Type: \tab Package\cr
Version: \tab 0.1 \cr
Date: \tab 2015-04-20\cr
License: \tab ? \cr
}
}
\author{


Jonas Haslbeck

Maintainer: <jonashaslbeck@gmail.com>
}
\references{
Loh, P. L., & Wainwright, M. J. (2013). Structure estimation for discrete graphical models: Generalized covariance matrices and their inverses. The Annals of Statistics, 41(6), 3022-3049.

Yang, E., Baker, Y., Ravikumar, P., Allen, G., & Liu, Z. (2014). Mixed graphical models via exponential families. In Proceedings of the Seventeenth International Conference on Artificial Intelligence and Statistics (pp. 1042-1050).

}
\keyword{ package }
\examples{}
